+++ src/extension.ts
@@ -5,9 +5,8 @@
 
 'use strict';
 
-import { registerAzureUtilsExtensionVariables } from '@microsoft/vscode-azext-azureutils';
-import { AzExtTreeDataProvider, AzureExtensionApiFactory, callWithTelemetryAndErrorHandling, createApiProvider, createAzExtOutputChannel, IActionContext, registerUIExtensionVariables } from '@microsoft/vscode-azext-utils';
-import type { AppResourceResolver } from '@microsoft/vscode-azext-utils/hostapi';
+import { registerAzureUtilsExtensionVariables, setupAzureLogger } from '@microsoft/vscode-azext-azureutils';
+import { AzExtTreeDataProvider, AzureExtensionApiFactory, callWithTelemetryAndErrorHandling, createApiProvider, createAzExtLogOutputChannel, IActionContext, registerUIExtensionVariables } from '@microsoft/vscode-azext-utils';
 import { apiUtils, GetApiOptions } from 'api/src/utils/apiUtils';
 import * as vscode from 'vscode';
 import { AzureResourcesApiInternal } from '../hostapi.v2.internal';
@@ -46,8 +45,9 @@ export async function activate(context: vscode.ExtensionContext, perfStats: { lo
 
     ext.context = context;
     ext.ignoreBundle = ignoreBundle;
-    ext.outputChannel = createAzExtOutputChannel('Azure Resource Groups', ext.prefix);
+    ext.outputChannel = createAzExtLogOutputChannel('Azure Resource Groups');
     context.subscriptions.push(ext.outputChannel);
+    context.subscriptions.push(setupAzureLogger(ext.outputChannel));
 
     registerUIExtensionVariables(ext);
     registerAzureUtilsExtensionVariables(ext);
@@ -64,7 +64,6 @@ export async function activate(context: vscode.ExtensionContext, perfStats: { lo
         activateContext.telemetry.properties.isActivationEvent = 'true';
         activateContext.telemetry.measurements.mainFileLoad = (perfStats.loadEndTime - perfStats.loadStartTime) / 1000;
 
-        setupEvents(context);
         ext.subscriptionProviderFactory = ext.isWeb ? createWebSubscriptionProviderFactory(context) : createAzureAccountSubscriptionProviderFactory();
 
         ext.tagFS = new TagFileSystem(ext.appResourceTree);
@@ -116,7 +115,7 @@ export async function activate(context: vscode.ExtensionContext, perfStats: { lo
 
     const v2ApiFactory: AzureExtensionApiFactory<AzureResourcesApiInternal> = {
         apiVersion: '2.0.0',
-        createApi: (options: GetApiOptions) => {
+        createApi: (options?: GetApiOptions) => {
             return createWrappedAzureResourcesExtensionApi(
                 {
                     apiVersion: '2.0.0',
@@ -132,7 +131,7 @@ export async function activate(context: vscode.ExtensionContext, perfStats: { lo
                         registerActivity
                     },
                 },
-                options.extensionId ?? 'unknown'
+                options?.extensionId ?? 'unknown'
             );
         }
     };
@@ -165,14 +164,4 @@ export async function activate(context: vscode.ExtensionContext, perfStats: { lo
 
 export function deactivate(): void {
     ext.diagnosticWatcher?.dispose();
-}
-
-function setupEvents(context: vscode.ExtensionContext): void {
-    // Event emitter for when a group is focused/unfocused
-    context.subscriptions.push(ext.emitters.onDidChangeFocusedGroup = new vscode.EventEmitter());
-    ext.events.onDidChangeFocusedGroup = ext.emitters.onDidChangeFocusedGroup.event;
-
-    // Event emitter for when an AppResourceResolver is registered
-    context.subscriptions.push(ext.emitters.onDidRegisterResolver = new vscode.EventEmitter<AppResourceResolver>());
-    ext.events.onDidRegisterResolver = ext.emitters.onDidRegisterResolver.event;
 }