+++ src/extensionVariables.ts
@@ -3,25 +3,15 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AzExtTreeDataProvider, IAzExtOutputChannel } from "@microsoft/vscode-azext-utils";
-import { AppResourceResolver } from "@microsoft/vscode-azext-utils/hostapi";
-import { DiagnosticCollection, Disposable, Event, EventEmitter, ExtensionContext, TreeView, UIKind, env } from "vscode";
+import { AzExtTreeDataProvider, IAzExtLogOutputChannel } from "@microsoft/vscode-azext-utils";
+import { DiagnosticCollection, Disposable, env, ExtensionContext, TreeView, UIKind } from "vscode";
 import { AzureResourcesApiInternal } from "../hostapi.v2.internal";
 import { ActivityLogTreeItem } from "./activityLog/ActivityLogsTreeItem";
 import { TagFileSystem } from "./commands/tags/TagFileSystem";
+import { AzureResourcesServiceFactory } from "./services/AzureResourcesService";
 import { AzureSubscriptionProvider } from "./services/SubscriptionProvider";
 import { ResourceGroupsItem } from "./tree/ResourceGroupsItem";
 import { TreeItemStateStore } from "./tree/TreeItemState";
-
-namespace extEmitters {
-    export let onDidChangeFocusedGroup: EventEmitter<void>;
-    export let onDidRegisterResolver: EventEmitter<AppResourceResolver>;
-}
-
-namespace extEvents {
-    export let onDidChangeFocusedGroup: Event<void>;
-    export let onDidRegisterResolver: Event<AppResourceResolver>;
-}
 
 export namespace extActions {
     export let refreshWorkspaceTree: (data?: ResourceGroupsItem | ResourceGroupsItem[] | null | undefined | void) => void;
@@ -42,16 +32,13 @@ export namespace ext {
     export let activityLogTree: AzExtTreeDataProvider;
     export let activityLogTreeItem: ActivityLogTreeItem;
     export let helpTree: AzExtTreeDataProvider;
-    export let outputChannel: IAzExtOutputChannel;
+    export let outputChannel: IAzExtLogOutputChannel;
     export let ignoreBundle: boolean | undefined;
     export const prefix: string = 'azureResourceGroups';
 
     export let tagFS: TagFileSystem;
     export let diagnosticWatcher: Disposable | undefined;
     export let diagnosticCollection: DiagnosticCollection;
-
-    export const emitters = extEmitters;
-    export const events = extEvents;
 
     export let azureTreeState: TreeItemStateStore;
 
@@ -63,6 +50,11 @@ export namespace ext {
 
     export namespace v2 {
         export let api: AzureResourcesApiInternal;
+    }
+
+    export namespace testing {
+        export let overrideAzureServiceFactory: AzureResourcesServiceFactory | undefined;
+        export let overrideAzureSubscriptionProvider: (() => AzureSubscriptionProvider) | undefined;
     }
 
     export const actions = extActions;